pyproject.toml
[project]
name = "llm-bench"
version = "0.1.0"
description = "Benchmark LLM inference backends (mlx-lm, ollama, llama.cpp, vLLM) on Apple Silicon"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "plotly>=5.18",
    "psutil>=5.9",
]

[project.optional-dependencies]
mlx = ["mlx-lm>=0.19"]
ollama = ["ollama>=0.4"]
llamacpp = ["llama-cpp-python>=0.3"]
vllm = ["vllm-metal>=0.1"]
all = ["llm-bench[mlx,ollama,llamacpp]"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[project.scripts]
llm-bench = "llm_bench.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
